show result of report current focus, navigator object, and selection in braille (#15858)
fixes #15844

Summary of the issue:
The report current focus and navigator object commands reports more information then NVDA normaly does by navigating to an object. Because of this, A user that only uses braille is unable to get this information. They are also unable to see acuratly what is selected if the selection is grater then one line, because when tethered to review, the selection information for text is lost, and when tethered to focus, the text is un selected when moving up or down a line.

Description of user facing changes
A user that only uses braille will be able to get the information described above

Description of development approach
I changed both the report focus and navigator object scripts, so that instead of calling speakObject, they call getObjectSpeech and saves its return value.
It then uses speech.speech.speak to speak this information
I then go over the list and remove every item that isn't a string, and then joins it with ' '.join
I then show this string in braille
In the reportCurrentSelection script, I get the information spoken using the _getSelectionMessageSpeech function, then show it in braille