No longer focus controls in settings dialog when they're invalid (#15897)
Related to #15894

Summary of the issue:
PR #15873 changed the validation for settings, so that when given control is invalid focus is moved to it. Unfortunately this works only when user didn't switch to a different panel. If they did the correct control was announced in speech, but the old panel is shown on screen.

Description of user facing changes
Invalid controls are no longer focused. This is still an improvement compared to 2023.3, where if one control was invalid the entire settings dialog was closed, thereby discarding all changes performed by the user.

Description of development approach
Removes calls to SetFocus from the validation code.