Fix hovering on cell in braille viewer (#15899)
Closes #15895

Summary of the issue:
Activation of routing cursor in the braille viewer by hovering the mouse on the cell does not work anymore in last alpha. This is due to wxPython upgrade, since int types is now expected for some functions when float was tolerated before. More specifically, wx.Colour now expects 3 integers.

Description of user facing changes
Mouse hovering to activate the routing cursor of a cell in the braille viewer works again.

Description of development approach
Convert the 3 numeric values passed to wx.Colour to integers.