Fix up of #15867 and other issues found during its investigation (#15906)
I have put various little changes in the same PR. Let me know if I should divide it.

Link to issue number:
Follow-up of #15867.

Summary of the issue:
Various issues seen in #15867 or while investigating on it (e.g. looking at usages of gesture.send()):

Script BrowseModeTreeInterceptor.script_passThrough in browseMode.py has a description. It should not since: 1. It is only dedicated to specific key which should not be remapped; 2. There should not be any input help associated to it, since this "pass through" should be transparent to users.
In Eclipse, when no selected auto-completion item can be found, NVDA+D should not pass the gesture through. Indeed gesture.send() should not be used since NVDA+D is not a native gesture of this IDE. An error message should be reported instead.
The PowerPoint script does not show up in its category, whereas it was the goal in Adding command gesture categories for appmodules #15867.
With 15867, scripts to set headers of rows or columns in Excel or Word have their description shortened and the details of its usage (e.g. 1 press, 2 presses) is not reported anymore in input help. This is not consistent with other scripts. It's better to keep more details in input help as for other scripts.