Remove a demo script (#15904)
Link to issue number:
Closes #11465

Summary of the issue:
NVDAObjects\IAccessible\winword.py contains a script that is not documented (no input help, not mentioned in user Guide). It has been introduced in commit 465dce8 when #3110 has been implemented as a demo script to test the new headers reading capability.

This script cause an error in the log when called outside of a table.

Description of user facing changes
No more error when pressing NVDA+shift+H in Word document.

Description of development approach
Just removed the script.