Honour choice when discarding the installation of an incompatible add-on (#15936)
Issue probably introduced in #15623.

Summary of the issue:
When installing an incompatible add-on (e.g. add-on last tested with NVDA 2023.3) from Windows Explorer, we get a warning asking to confirm the installlation. If I answer No, the installation process still continues.
It's because the function to display the dialog box now returns a 2-tuple which always evals to True.

Description of user facing changes
Answering No in the incompatibility warning dialog will now stop the installation process as expected.

Description of development approach
Only check the first element of the tuple to know the choice made in the dialog.