Ensure the mouse does not play audio coordinates in sleep mode (#15962)
Fixes #8059

Summary of the issue:
This issue seems to be because in the core pump, mouse movement events are being handled, but when executing the mouse movement event, there is no check for whether it is in sleep mode. And the function to play audio coordinates is executed accordingly.

Description of user facing changes
The audio coordinates are no longer played when the mouse moves over an application in sleep mode.

Description of development approach
Simultaneously determine whether to report audio coordinates and whether to be in sleep mode.