Remove and ignore  English nvda.po as it is not needed, yet is provided by Crowdin. (#15989)
Mentioned in pr #15968

Summary of the issue:
When exporting translated po files from Crowdin, English is also included. this is not needed and just takes up space in the repository.

Description of user facing changes
None

Description of development approach
Remove the accidentally committed English nvda.po
Add English nvda.po to .gitignore so that it is not committed again.