Properly destroy Add-on Store dialog when it is closed (#16019)
Related to #15041 and #15105

Summary of the issue:
Ever since Add-on Store was introduced NVDA was intermittently failing to restart on my machine. This occurred only when I've opened Add-on Store before restarting. After debugging it turned out the following happens:

When exiting NVDA had to destroy the store dialog, as it was kept alive due to circular reference
The new instance of NVDA tried to wait for the old instance to exit, but since the exit process took extra long due to destroying the dialog the new copy gave up and didn't start
Description of user facing changes
NVDA should no longer fail to restart intermittently after opening and closing Add-on Store.

Description of development approach
The reference to one of BoxSizerHelper objects was stored on the store dialog. Since each BoxSizerHelper also stores reference to its parent the store dialog could not be destroyed properly when closing. This circular reference cycle was broken by no longer storing the BoxSizerHelper on the instance - it is either used as a local variable, or passed to the method which needs it as an parameter.

Testing strategy:
Ensured store dialog is still functioning
Stress tested restarting NVDA on a machine where this was failing and verified it restarts successfully every single time
Ensured no instances are stored in gui.settingsDialogs.SettingsDialog._instances when the Add-on Store dialog is closed
Known issues with pull request:
Storing a BoxSizerHelper as a dialog member is an easy mistake to make in the future. I will submit a PR which makes it less likely to occur by storing a weak reference to the parent in the helper instance. Since this is a bigger and more risky change it will target master
The approach taken by Add-on Store: prevent duplicate instances from opening #15105 i.e. ignoring the warning about existing instances is sub-optimal. As demonstrated by this PR ignoring these alive references may cause issues in unrelated areas of NVDA. In my opinion we should restore the old behavior i.e. live references should once again cause a warning. If this is fine I'll submit a PR to that effect against master.