Adobe Reader: support alternative text on formulas in PDFs (#16067)
Fixes #12715

Summary of the issue:
Authors of PDFs can provide alternative text on formular nodes via the 'alt' attribute. However, NvDA ignores this and replaces it with a single space, with the assumption that the formula will contain an inner mathml tree structure.
The idea of embedding a full mathml tree structure with in formula nodes was never standardized, and only a few experimental PDFs actually exist. However, it is becoming common for authors to place alt text (perhaps a simple english representation of math) on formula nodes.

Description of user facing changes
In Adobe Reader, alternative text for formulas will be reported if provided by the author.

Description of development approach
Adobe Acrobat vbuf backend: Don't replace content in formula tags with a space, rather only do this for math nodes. this ensures NvDA still exposes alt text on formulas. this approach still allows for supporting embedded mathml trees if they exist as a direct child of the formula.