No longer store strong references to parent object in guiHelper.BoxSizerHelper to decrease likelihood of circular references in NVDA's GUI (#16079)
Related to #16019

Summary of the issue:
BoxSizerHelper class stores a strong reference to its parent. When it is mistakenly assigned to a member of its parent a circular reference is created making it impossible for Python's garbage collector to destroy the dialog when it goes out of scope.

Description of user facing changes
Should not be noticeable.

Description of development approach
BoxSizerHelper now stores a weak reference to its parent.