Fix exclusion in Flake8 configuration (#16087)
Related to #16071

Summary of the issue:
When linting we exclude certain directories where the code is auto generated. These exclusions are not working (this probably regressed with the Flake8 update, though I haven't checked).

Description of user facing changes
When linting exclusions in Flake8 configuration are once again respected.

Description of development approach
Flake8 considers that paths in the exclusion list are provided relative to the config file location not to the CWD, our exclusions were modified to account for this
After the file was modified I started getting errors due to usage of inline comments, apparently this was never supposed to work, as per this quote from the documentation:
Following the recommended settings for Python’s configparser, Flake8 does not support inline comments for any of the keys. So while this is fine:...

Therefore we no longer use inline comments in the config.