Report the code of combined characters (#16098)
Closes #5011.

Summary of the issue:
Some characters are built with two or more Unicode characters: the first is a normal character and the following ones are combining characters. For example the character "é". It is built from the character "e" and the character "combining acute accent".

Pressing numpad2 twice, the description of the two characters are reported. However pressing numpad2 3 times, no character code is reported and the compound character is just reported as at first press; thus, triple press brings no useful information in this case.

Description of user facing changes
When pressing three times numpad2:

if the character under the review cursor is a single character, its decimal and hexadecimal codes will be reported as in pevious NVDA versions
if the character under the review cursor is compound of two or more characters, the codes for each character building it will be reported:
In braille, both decimal and hexadecimal codes will be reported for each character
With speech, only the decimal code is reported to avoid cluttering the speech output with too much numeric information that may be hard to parse.