Fix occasional error when the session is locked/unlocked (#16121)
Fix #16120
Related to #15400.

Summary of the issue:
Sporadic errors still occur when locking/unlocking Windows session. As per #15400, this is due to the fact that api.getFocusObject returns an NVDAObjects.NVDAObject but we use it as it was an NVDAObjects.window.Window. But that's not the case when the focus is (or was) winAPI.secureDesktop._handleSecureDesktopChange._SecureDesktopNVDAObject.

Description of user facing changes
Probably none; these errors did not seem to have any visible impact.

Description of development approach
Check the class of the object returned by api.getFocusObject before using its properties.