Ignore padding dots at low symbol reporting levels (#16141)
Closes #15845

Summary of the issue:
Padding dots in table of contents are reported even at low punctuation levels.

Description of user facing changes
Padding dots are not reported anymore at low punctuation levels.

Description of development approach
In symbol file, define a complex symbol to identify padding dots as 4 or more dots. Assign level to "all", e.g. as "end of sentence dot". And define "send real symbol to synthesizer" to "always" so that a pause is kept between the text before the dots and the text after them.

In character processing, change the order of symbol processing as follows:

complex symbols rules
repetition rules
simple symbol rules
Before, repetition rule was the first. This has been done so that the repetition rule do not override the new rule for the padding dots complex symbol.
In any case, I do not think that there was any use case of the repetition rule being used with complex symbols.