Use the add-on's (translated) display name in the removal confirmation dialog (#16157)
Summary of the issue:
When removing an add-on, there is a confirmation dialog.
The add-on's name used in this dialog is NVDA's internal name (formerly "id"). It should be the name displayed in the GUI instead; this name displayed in the GUI is also translated whereas the internal name is not.

Description of user facing changes
The translated full name will be used in the add-on removal confirmation dialog, instead of the internal name.