Fix text paragraph navigation speech (#16145)
Closes #16143

Summary of the issue:
NVDA reports roles incorrectly when navigating using P quickNav

Description of user facing changes
Fixed speech so that p quickNav command behaves as expected.

Description of development approach
A quick investigation shows that this is the behavior of speech.speakTextInfo(info, reason=OutputReason.QUICKNAV) as called in browseMode.py:216. I didn't change that in textNav PR.

This makes sense for other QuickNav commands. For example on that page if you press B, you get:

Open global navigation menu Button
Note that the word button goes in the end. If you stumble upon the same button via Control+Up/Down then the speech is

Button Open global navigation menu
So since text paragraph navigation is more similar to caret navigation in the sense that it navigates to text that might contain different roles, I propose to change to reason=OutputReason.CARET for text paragraph navigation as it seems to fix this issue.