Make installer process safer (#16174)
Fixes #16122

Summary of the issue:
NVDA process sometimes fails to quit correctly (see #16123).
When running the installer, this causes the installer to fail partway through.
If the installer fails partway through, the installation may end up in an unrecoverable state.

For certain files, NVDA requests windows to move or delete these on reboot, if the move/delete fails initially.
However, if the Windows API request fails, NVDA does not handle the failure the same way as it handles other failures - raising a retryable failure and reverting the change if possible.

Description of user facing changes
The NVDA installation process should be more resilient if the previous process is still running.

Description of development approach
NVDA attempts to remove all known exes in a batch as the first operation in the installation. If any known NVDA exe is running, the installation is aborted with a retry able failure and all exes returned to their original state. This means the user can attempt the installation again - i.e. waiting a bit longer for the process to hopefully exit.

If NVDA fails to request windows to remove certain files on restart, this is handled the same as if the files are failed to be removed.

More logging has been added and improved.