Make figure reporting configureable, and figures quick navigable (#16176)
fixes #10826
Fixes #14349

Summary of the issue:
It has hitherto not been possible to navigate by figure with quick navigation keys, or to disable the reporting of figures and captions.

Description of user facing changes
Added a script to jump to next/previous figure in browse mode. Added an option to disable the reporting of figures and captions, and a script to toggle this setting.

Description of development approach
Examined similar code and based new code on existing code.